+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - suppress mermaid rendering errors
 - allow reverting untracked (new) files in updated files sidebar
 - improved process killing reliability in bash tool
+- added openUrl functionality to Extension context to open URLs in a new window, external or in a modal overlay
 
 ## [0.56.0]